Dashboard

+++ OdontoTech/DataAccessLayer/BairroDAL.cs
@@ -13,12 +13,10 @@ namespace DataAccessLayer
     {
         MySqlConnection conn = new MySqlConnection(DBConfig.CONNECTION_STRING);
         MySqlCommand cmd = new MySqlCommand();
-
-        public string Inserir(Bairro bairro)
+        public string Insert(Bairro bairro)
         {
             cmd.Connection = conn;
-            cmd.CommandText = $"INSERT INTO bairro (nomeBairro,idCidade) values (@nomeBairro,@idCidade)";
-
+            cmd.CommandText = "INSERT INTO bairro(nomeBairro, idCidade) values(@nomeBairro, @idCidade)";
             cmd.Parameters.AddWithValue("@nomeBairro", bairro.Nome);
             cmd.Parameters.AddWithValue("@idCidade", bairro.Cidade.Id);
 
@@ -45,8 +43,13 @@ namespace DataAccessLayer
                 conn.Dispose();
             }
         }
-        public string Deletar(Bairro bairro)
+        public string Delete(Bairro bairro)
         {
+            if (bairro.Id == 0)
+            {
+                return "Bairro informada inválido!";
+            }
+
             cmd.Connection = conn;
             cmd.CommandText = "DELETE FROM bairro WHERE idBairro = @ID";
             cmd.Parameters.AddWithValue("@ID", bairro.Id);
@@ -66,12 +69,13 @@ namespace DataAccessLayer
                 conn.Dispose();
             }
         }
-        public string Atualizar(Bairro bairro)
+        public string Update(Bairro bairro)
         {
             cmd.Connection = conn;
-            cmd.CommandText = "UPDATE bairro SET nomeBairro = @nomeBairro WHERE idBairro = @idBairro";
-            cmd.Parameters.AddWithValue("@idBairro", bairro.Id);
+            cmd.CommandText = "UPDATE bairro SET nomeBairro = @nomeBairro, idCidade = @idCidade WHERE idBairro = @idBairro";
             cmd.Parameters.AddWithValue("@nomeBairro", bairro.Nome);
+            cmd.Parameters.AddWithValue("@idCidade", bairro.Cidade.Id);
+            cmd.Parameters.AddWithValue("@idBairro", bairro.Id);
 
             try
             {
@@ -89,25 +93,28 @@ namespace DataAccessLayer
                 conn.Dispose();
             }
         }
-        public List<Bairro> SelecionaTodos()
+        public List<Bairro> GetAll()
         {
             cmd.Connection = conn;
             cmd.CommandText = "SELECT * FROM bairro";
+
             try
             {
                 conn.Open();
                 MySqlDataReader reader = cmd.ExecuteReader();
-                List<Bairro> Bairros = new List<Bairro>();
+                List<Bairro> bairros = new List<Bairro>();
+
                 while (reader.Read())
                 {
                     Bairro temp = new Bairro();
-
                     temp.Id = Convert.ToInt32(reader["idBairro"]);
                     temp.Nome = Convert.ToString(reader["nomeBairro"]);
                     temp.Cidade.Id = Convert.ToInt32(reader["idCidade"]);
-                    Bairros.Add(temp);
+
+                    bairros.Add(temp);
                 }
-                return Bairros;
+
+                return bairros;
             }
             catch (Exception)
             {
@@ -118,28 +125,26 @@ namespace DataAccessLayer
                 conn.Dispose();
             }
         }
-        public Bairro GetByID(int id)
+        public Bairro GetById(int idBairro)
         {
             cmd.Connection = conn;
-            cmd.CommandText = "SELECT * FROM bairro WHERE idBairro = @ID";
-            cmd.Parameters.AddWithValue("@ID", id);
+            cmd.CommandText = "SELECT * FROM bairro WHERE idBairro = @idBairro";
+            cmd.Parameters.AddWithValue("@idBairro", idBairro);
 
             try
             {
                 conn.Open();
                 MySqlDataReader reader = cmd.ExecuteReader();
-                Bairro temp = new Bairro();
+                Bairro bairro = new Bairro();
 
                 while (reader.Read())
                 {
-
-                    temp.Id = Convert.ToInt32(reader["idBairro"]);
-                    temp.Nome = Convert.ToString(reader["nomeBairro"]);
-                    temp.Cidade.Id = Convert.ToInt32(reader["idCidade"]);
-
-
+                    bairro.Id = Convert.ToInt32(reader["idBairro"]);
+                    bairro.Nome = Convert.ToString(reader["nomeBairro"]);
+                    bairro.Cidade.Id = Convert.ToInt32(reader["idCidade"]);
                 }
-                return temp;
+
+                return bairro;
             }
             catch (Exception)
             {
@@ -165,17 +170,46 @@ namespace DataAccessLayer
 
                 while (reader.Read())
                 {
+                    bairro.Id = Convert.ToInt32(reader["idBairro"]);
+                    bairro.Nome = Convert.ToString(reader["nomeBairro"]);
+                    bairro.Cidade.Id = Convert.ToInt32(reader["idCidade"]);
+                }
+
+                return bairro;
+            }
+            catch (Exception)
+            {
+                throw new Exception("Erro no Banco de dados.Contate o administrador.");
+            }
+            finally
+            {
+                conn.Dispose();
+            }
+        }
+
+        public List<Bairro> GetByCidade(Cidade cidade)
+        {
+            cmd.Connection = conn;
+            cmd.CommandText = "SELECT * FROM bairro WHERE idCidade = @idCidade";
+            cmd.Parameters.AddWithValue("@idCidade", cidade.Id);
+
+            try
+            {
+                conn.Open();
+                MySqlDataReader reader = cmd.ExecuteReader();
+                List<Bairro> bairros = new List<Bairro>();
+
+                while (reader.Read())
+                {
                     Bairro temp = new Bairro();
-
-
                     temp.Id = Convert.ToInt32(reader["idBairro"]);
                     temp.Nome = Convert.ToString(reader["nomeBairro"]);
                     temp.Cidade.Id = Convert.ToInt32(reader["idCidade"]);
 
-                    bairro = temp;
+                    bairros.Add(temp);
                 }
 
-                return bairro;
+                return bairros;
             }
             catch (Exception)
             {
